subgraph connectivity by pathway type
a pathway type can be specified by (source-type-specifiers, target-type-specifiers)
For example we can have (layer, [layer, mtype])
This will generate the edge types by crossing the possible values of source types and target types
as specified.
For each edge type, source and target nodes will be chosen.
Note that our definitions allows for overlapping sources and targets.
The resulting matrix will be size N with N equal to the number of nodes
in the union of sources and targets.
The resulting matrix will keep the connections that go only from the selected sources to targets.